% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ve.R
\name{add_ve}
\alias{add_ve}
\title{Calculate VE based on psi_bar_0 and psi_bar_1}
\usage{
add_ve(psi_bar)
}
\arguments{
\item{psi_bar}{{A named numeric vector containing point estimates for
\verb{psi_bar_0, psi_bar_1}}}
}
\value{
{A named numeric vector containing point estimates for
\verb{psi_bar_0, psi_bar_1}, and \code{ve}}
}
\description{
This function computes VE as the proportion reduction in risk due to
vaccination (ie. VE = 1 - psi_bar_0 / psi_bar_1)
}
