% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matched_adata.R
\name{clean_matched_data}
\alias{clean_matched_data}
\title{Get analysis matched dataset from matched cohort}
\usage{
clean_matched_data(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  tau,
  pair_censoring = TRUE
)
}
\arguments{
\item{matched_data}{A data frame representing a matched cohort}

\item{outcome_name}{Character string representing the original outcome variable in  \code{matched_data}}

\item{event_name}{Character string representing the original event variable in \code{matched_data}}

\item{trt_name}{Character string representing the original vaccination status variable in \code{matched_data}}

\item{time_name}{Character string representing the original vaccination time variable in \code{matched_data}}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}
}
\value{
A data frame representing the analysis-eligible matched dataset (a subset of \verb{matched data}).
Contains all variables in \code{matched_data}, plus the following variables
\describe{
\item{<event_name>_d}{Event variable after adjusting for pair_censoring}
\item{<outcome_name>_matched}{Outcome variable after adjusting for pair_censoring}
\item{T_d}{Time to endpoint from matched index date (adjusted for pair_censoring if requested)}
}
}
\description{
This function modifies the original matched data, preparing it for use in
analysis. Namely, this includes
\itemize{
\item (if requested) censoring matched pairs at the time
of the unvaccinated individual's vaccination
\item creating the outcome time from matching index date to endpoint
\item restricting to matched  pairs in which both individuals are eligible \code{tau} days after the matching
index date
}
}
