% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\name{compute_boot_ci}
\alias{compute_boot_ci}
\alias{compute_wald_ci}
\alias{compute_percentile_ci}
\title{Compute Wald or percentile bootstrapped confidence interval}
\usage{
compute_boot_ci(
  x,
  boot_x,
  ci_type,
  alpha = 0.05,
  transform = NULL,
  z_star = NULL
)

compute_wald_ci(x, boot_x, alpha = 0.05, transform, z_star = NULL)

compute_percentile_ci(boot_x, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector of point estimates}

\item{boot_x}{A matrix of bootstrapped estimates where the rows are the
bootstrap iterations and the columns are the time points of interest.}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{transform}{If \code{ci_type = "wald"}, a character string indicating the scale on which
to compute Wald confidence intervals that are then transformed back to the original scale.
Options are \code{logit} for the transformation log(x/(1-x)) or \code{log_ve} for the transformation
log(1-x).}

\item{z_star}{If \code{ci_type = "wald"}, a specific critical value used to
to compute Wald confidence intervals (assumed to be positive). If used, \code{alpha} argument is ignored.}
}
\value{
A matrix containing the lower and upper confidence intervals and
relevant bootstrap standard errors.
}
\description{
Internally calls \code{\link[=compute_wald_ci]{compute_wald_ci()}} and/or \code{\link[=compute_percentile_ci]{compute_percentile_ci()}}
depending on the value of \code{ci_type}
}
