% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_marginal_ve.R
\name{compute_marginal_ve}
\alias{compute_marginal_ve}
\title{Get the marginal cumulative incidence in the treated and untreated
groups based on a Cox model or Kaplan Meier curve.}
\usage{
compute_marginal_ve(
  adata,
  adata_outcome_name,
  adata_event_name,
  adata_trt_name,
  times,
  method,
  adjust = NULL,
  censor_time = max(times),
  separate = TRUE,
  return_models = TRUE
)
}
\arguments{
\item{adata}{A data frame that represents the analysis data set of a clinical trial.}

\item{adata_outcome_name}{Character string specifying the time to event variable in \code{adata}. The
time should be the time to event from vaccination/matched index date}

\item{adata_event_name}{Character string specifying the event variable in \code{adata}}

\item{adata_trt_name}{Character string specifying the treatment variable in \code{adata}}

\item{method}{Character string specifying method for survival estimation ("cox" for
Cox proportional hazards regression model or "km" for Kaplan-Meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{separate}{If \code{method = cox} and \code{separate = TRUE}, Cox regression models are fit in the treated and untreated groups separately.}

\item{return_models}{Logical value: Should fitted survival models be returned?}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A matrix of estimates. The columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.}
\item{models}{If \code{return_models = TRUE}, the models used to compute risk and VE}
}
}
\description{
#' @description
Internally calls \code{\link[=compute_cox_ve]{compute_cox_ve()}} or \code{\link[=compute_km_ve]{compute_km_ve()}} depending
on the value of \code{method}
}
