% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_bar.R
\name{compute_psi_bar}
\alias{compute_psi_bar}
\title{Marginalize conditional risks to obtain overall cumulative incidences}
\usage{
compute_psi_bar(psi_d, gp_list)
}
\arguments{
\item{psi_d}{A data frame with conditional risk predictions containing:
\itemize{
\item \code{cuminc_0}: cumulative incidence under no vaccine,\eqn{\psi_0(t_0; d, x)}
\item \code{cuminc_1}: cumulative incidence  under vaccine,\eqn{\psi_1(t_0; d, x)}
\item \verb{<adjust_vars>}: all covariate columns
\item \verb{<time_name>}: vaccination day column
Typically output from \code{\link[=compute_psi_d]{compute_psi_d()}}.
}}

\item{gp_list}{A list containing the marginalizing distributions
#TODO: assumptions on gp_list
must have group_name}
}
\value{
A named numeric vector with two elements:
\itemize{
\item \code{cuminc_0}: Overall cumulative incidence under no vaccine
\item \code{cuminc_1}: Overall cumulative incidence under vaccine
}
}
\description{
Takes day- and covariate-specific risks \eqn{\psi_0(t_0; d, x)} and \eqn{\psi_1(t_0; d, x)}
and averages them using weights from \code{gp_list} to produce overall
cumulative incidences under each vaccine regime
}
\details{
The function performs a weighted average in two stages:
\enumerate{
\item Within each covariate group x, average over vaccination days d using g(d|x)
\item Across covariate groups, average using p(x)
}
}
\seealso{
\itemize{
\item \code{\link[=compute_psi_d]{compute_psi_d()}} for generating \code{psi_d} input
\item \code{\link[=get_observed_gp]{get_observed_gp()}} or \code{\link[=get_matching_gp]{get_matching_gp()}} for creating \code{gp_list}
}
}
