% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_bar.R
\name{compute_psi_bar_t0}
\alias{compute_psi_bar_t0}
\title{Compute overall cumulative incidences at a single timepoint}
\usage{
compute_psi_bar_t0(fit_0, fit_1, exposure_time, t0, tau, newdata, gp_list)
}
\arguments{
\item{fit_0}{A fitted model returned from \code{\link[=fit_model_0]{fit_model_0()}}}

\item{fit_1}{A fitted model returned from \code{\link[=fit_model_1]{fit_model_1()}}}

\item{exposure_time}{Name of the time-to-exposure variable  in \code{newdata}.
Used to compute \eqn{\psi_0(t_0; d,x)} where \eqn{d + \tau} and \eqn{d + t_0} are needed.}

\item{t0}{Time since exposure at which to evaluate cumulative incidence.}

\item{tau}{Delay period}

\item{newdata}{New data at which to do predictions.}

\item{gp_list}{A list with two data frames:
\itemize{
\item \strong{g_weights} Data frame of covariate-conditional exposure-day probabilities \eqn{g(d \mid x)}.
Must include:
\itemize{
\item \code{group_id}: covariate group identifier
\item \verb{<exposure_time>}: exposure time variable
\item \code{prob_g}: probability of exposure time given the covariates
\item all variables in \code{covariates}
}
\item \strong{p_weights} Data frame of covariate probabilities \eqn{p(x)}.
Must include:
\item \code{group_id}: covariate group identifier
\item \code{prob_p}: marginal probability of each covariate group
\item all variables in \code{covariates}
}

Default is \code{NULL} in which case each row of \code{psi_dx} gets equal weight.}
}
\value{
Named numeric vector with \code{cuminc_0}, \code{cuminc_1}
}
\description{
Compute overall cumulative incidences at a single timepoint
}
\keyword{internal}
