% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_bar.R
\name{compute_psi_bar_times}
\alias{compute_psi_bar_times}
\title{Compute overall cumulative incidences at multiple timepoints}
\usage{
compute_psi_bar_times(
  fit_0,
  fit_1,
  exposure_time,
  eval_times,
  tau,
  newdata,
  gp_list
)
}
\arguments{
\item{fit_0}{A fitted model returned from \code{\link[=fit_model_0]{fit_model_0()}}}

\item{fit_1}{A fitted model returned from \code{\link[=fit_model_1]{fit_model_1()}}}

\item{exposure_time}{Name of the time-to-exposure variable  in \code{newdata}.
Used to compute \eqn{\psi_0(t_0; d,x)} where \eqn{d + \tau} and \eqn{d + t_0} are needed.}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure. All values must be greater
than \code{tau} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (tau+1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{tau}{Delay period}

\item{newdata}{New data at which to do predictions.}

\item{gp_list}{A list with two data frames:
\itemize{
\item \strong{g_weights} Data frame of covariate-conditional exposure-day probabilities \eqn{g(d \mid x)}.
Must include:
\itemize{
\item \code{group_id}: covariate group identifier
\item \verb{<exposure_time>}: exposure time variable
\item \code{prob_g}: probability of exposure time given the covariates
\item all variables in \code{covariates}
}
\item \strong{p_weights} Data frame of covariate probabilities \eqn{p(x)}.
Must include:
\item \code{group_id}: covariate group identifier
\item \code{prob_p}: marginal probability of each covariate group
\item all variables in \code{covariates}
}

Default is \code{NULL} in which case each row of \code{psi_dx} gets equal weight.}
}
\value{
A matrix of estimates where the columns are the terms \code{cuminc_0} and \code{cuminc_1},
and the rows are the time points of interest.
}
\description{
Wrapper that computes cumulative incidences at multiple
timepoints. Calls \code{compute_psi_bar_t0()} internally for each timepoint.
Models are fitted once before calling this function to allow efficient evaluation
at multiple timepoints without refitting.
}
\details{
This function expects models already fitted via \code{\link[=fit_model_0]{fit_model_0()}} and
\code{\link[=fit_model_1]{fit_model_1()}}. It performs G-computation by predicting and marginalizing over
conditional risks.
}
