% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_dx.R
\name{compute_psi_dx_t0}
\alias{compute_psi_dx_t0}
\title{Compute day- and covariate- specific cumulative incidences}
\usage{
compute_psi_dx_t0(fit_0, fit_1, exposure_time, t0, tau, newdata)
}
\arguments{
\item{fit_0}{A fitted model returned from \code{\link[=fit_model_0]{fit_model_0()}}}

\item{fit_1}{A fitted model returned from \code{\link[=fit_model_1]{fit_model_1()}}}

\item{exposure_time}{Name of the time-to-exposure variable  in \code{newdata}.
Used to compute \eqn{\psi_0(t_0; d,x)} where \eqn{d + \tau} and \eqn{d + t_0} are needed.}

\item{t0}{Time since exposure at which to evaluate cumulative incidence.}

\item{tau}{Delay period}

\item{newdata}{New data at which to do predictions.}
}
\value{
A data frame with one row per row of \code{newdata} with the predicted
cumulative incidences and component survival probabilities:
\itemize{
\item \code{psi_0_dx}, \code{surv_0_d_plus_tau}, \code{surv_0_d_plus_t0}
\item \code{psi_1_dx}, \code{surv_1_tau}, \code{surv_1_t0}
}
}
\description{
Wrapper that calls internal functions for predicting cumulative incidences from
fitted hazard models. Returns the predicted exposure-specific cumulative
incidences side by side for each time- and covariate- pair in \code{newdata}.
}
\details{
Definitions of the cumulative incidences returned:
\deqn{\psi_0(t_0; d,x) = 1 - S_0(d+t_0; x)\,/\,S_0(d+\tau; x)}
\deqn{\psi_1(t_0; d,x) = 1 - S_1(t_0; d,x)\,/\,S_1(\tau; d,x)}

where \eqn{d} represents exposure time, \eqn{x} represents covariates, and
\eqn{S_v} represents the survival probability from hazard model for exposure \eqn{v}.
}
\seealso{
\code{\link[=predict_from_model_0]{predict_from_model_0()}}, \code{\link[=predict_from_model_1]{predict_from_model_1()}}
}
