% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ve.R
\name{compute_ve}
\alias{compute_ve}
\title{Compute proposed causal VE estimand given the fitted survival models at various
time points of interest}
\usage{
compute_ve(fit_0, fit_1, time_name, times, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted survival model for the unvaccinated group}

\item{fit_1}{A fitted survival model for the vaccinated group}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{gp_list}{List of marginalizing distributions. Used to determine which x, d
will be used for prediction}
}
\value{
A matrix of estimates where the columns are the terms \verb{risk_0, risk_1}, and \code{ve}
and the rows are the time points of interest.
}
\description{
\code{compute_ve()} returns the VE estimates at multiple time points. This is function is a
wrapper around \code{compute_ve_t0()}.
}
