% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ci.R
\name{compute_wald_ci}
\alias{compute_wald_ci}
\alias{compute_percentile_ci}
\title{Compute bootstrapped Wald or percentile confidence intervals}
\usage{
compute_wald_ci(pt_est, boot_samples, alpha, z_star = NULL)

compute_percentile_ci(boot_samples, alpha)
}
\arguments{
\item{pt_est}{A named numeric vector containing point estimates for
\verb{psi_bar_0, psi_bar_1, ve}}

\item{boot_samples}{A matrix containing estimates from all bootstrap replications. Rows
represent bootstrap iterations. Columns are named \code{psi_bar_0}, \code{psi_bar_1} and \code{ve}.}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{z_star}{Numeric vector with same length as \code{pt_est}. Specify the critical value to use when computing the confidence
interval for each term in \code{pt_est}. Useful when computing simultaneous confidence intervals.}
}
\value{
Matrix containing the lower and upper bounds. For Wald confidence intervals,'
transformed standard errors are also included.
}
\description{
\code{compute_wald_ci()}  computes Wald-style bootstrapped confidence intervals for
\code{psi_bar_0}, \code{psi_bar_1} and \code{ve}. Confidence intervals for \verb{psi_bar_0, psi_bar_1}
are computed on the logodds scale (log(x/(1-x))) then transformed back to the
original scale. The confidence interval for \code{ve} is computed on the scale of \code{log(1-x)}
and then back-transformed.

\code{compute_percentile_ci()} computes percentile bootstrapped confidence intervals.
}
