% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\docType{data}
\name{.backward_transformations}
\alias{.backward_transformations}
\title{Internal backward (inverse) transformations}
\format{
An object of class \code{list} of length 3.
}
\usage{
.backward_transformations
}
\description{
These define how to back-transform point estimates
Functions take arguments \verb{(eta, sd, z)} and return values
on the original scale.
}
\keyword{internal}
