% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\docType{data}
\name{.forward_transformations}
\alias{.forward_transformations}
\title{Internal forward transformations}
\format{
An object of class \code{list} of length 3.
}
\usage{
.forward_transformations
}
\description{
These define how estimates are transformed before applying
the Wald CI (e.g., taking log or logit).
}
\keyword{internal}
