% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\docType{data}
\name{.term_transform_map}
\alias{.term_transform_map}
\title{Internal map of term → transformation type}
\format{
An object of class \code{character} of length 4.
}
\usage{
.term_transform_map
}
\description{
Internal map of term → transformation type
}
\keyword{internal}
