% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bootstrap_ci.R
\name{estimate_bootstrap_ci}
\alias{estimate_bootstrap_ci}
\title{Helper for constructing bootstrapped confidence intervals}
\usage{
estimate_bootstrap_ci(
  one_boot_function,
  one_boot_args,
  ci_type,
  boot_reps,
  pt_est = NULL,
  alpha = 0.05,
  keep_boot_samples = TRUE,
  n_cores = 1
)
}
\arguments{
\item{one_boot_function}{Function that computes one bootstrap iteration and
returns the bootstrap estimates}

\item{one_boot_args}{List of arguments to pass to \code{one_boot_function}}

\item{ci_type}{Method for constructing bootstrap confidence intervals. One of
\code{"wald"}, \code{"percentile"}, or \code{"both"}.
\itemize{
\item \code{"wald"} (default): Computes Wald-style intervals using bootstrap standard errors.
See \strong{Confidence intervals} section for details.
\item \code{"percentile"}: Computes percentile bootstrap intervals.
\item \code{"both"}: Computes and returns both sets of intervals.
}}

\item{boot_reps}{Number of bootstrap replicates for confidence intervals.
Recommended to use at least 1000 for publication-quality results. Use
smaller values (e.g., 10-100) for initial exploration. Default: \code{0} (no
bootstrapping).}

\item{pt_est}{A matrix of point estimates, with columns corresponding to
cumulative incidence and effect measures, and rows representing evaluation
timepoints. This argument must be provided when \code{ci_type = "wald"}. Ignored
when \code{ci_type = "percentile"}.}

\item{alpha}{Significance level for confidence intervals (Confidence level =
100*(1-\code{alpha})\%). Default: \code{0.05}.}

\item{keep_boot_samples}{Logical; return bootstrap samples? Default:
\code{TRUE}. Must be set to \code{TRUE} if user plans to use \code{\link[=add_simultaneous_ci]{add_simultaneous_ci()}}
to obtain simultaneous confidence intervals.}

\item{n_cores}{Integer; parallel cores for bootstrapping. Passed to
\code{parallel::mclapply} as \code{mc.cores}. On Unix-like OS only; not available on
Windows. Default: \code{1}.}
}
\value{
A list containing the following:
\describe{
\item{ci_estimates}{A list of matrices, one for each term, containing the
lower and upper confidence interval bounds}
\item{n_success_boot}{A numeric vector of the number of successful bootstrap
samples for each time point.(Success bootstrap samples are
those that result in non-missing valid point estimates.)}
\item{boot_samples}{If \code{keep_boot_samples = TRUE}, a list of matrices, one
for each term, containing the bootstrap estimates. Rows are the
bootstrap iterations and columns are the time points.}
}

When \code{boot_reps = 0}, returns \code{NULL}.
}
\description{
This function computes Wald and percentile bootstrapped
confidence intervals.
}
\keyword{internal}
