% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ci.R
\name{estimate_ci}
\alias{estimate_ci}
\title{Estimate bootstrapped confidence intervals for VE}
\usage{
estimate_ci(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  t0,
  censor_time,
  tau,
  boot_formula_0,
  boot_formula_1,
  matched_data,
  gp_list,
  ci_type,
  limit_type,
  n_boot,
  pt_est = NULL,
  alpha = 0.05,
  return_boot = TRUE
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{marginalizing_dist}{Character string describing the type of estimated
day/covariate distributions to marginalize over. Values are "observed" or
"matched". Alternatively, for pre-specified marginalizing distributions can
be a list of 2 dataframes named \code{g_dist} and \code{p_dist}.}

\item{t0}{The time (relative to time of vaccination) at which to return
cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
If interested in multiple timepoints, recommended to set this to the maximum timepoint of interest
and then using \code{\link[=timepoints]{timepoints()}}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{boot_formula_0}{A formula for estimating hazards in
unvaccinated group for bootstrap samples (not a model)}

\item{boot_formula_1}{A formula for estimating hazards in
vaccinated group for bootstrap samples (not a model)}

\item{matched_data}{A data frame representing the matched cohort}

\item{gp_list}{List of two dataframes named \code{g_dist} and \code{p_dist} representing
the marginalizing distributions to use.}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{limit_type}{Character string indicating whether the marginalizing
distributions of interest are the estimated distributions in the "limit"
or "fixed" sense. For a prespecified marginalizing_dist, limit_type is
always fixed.}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{pt_est}{A named numeric vector containing point estimates for
\verb{psi_bar_0, psi_bar_1, ve}}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}
}
\value{
A list containing the following:
\describe{
\item{ci_estimates}{A matrix containing the lower and upper confidence interval bounds
and bootstrapped standard eror. When \code{ci_type = "wald"}, the bootstrapped standard errors
on the transformed scale are also included.}
\item{n_success_boot}{The number of bootstrap samples used to compute confidence interval}
\item{boot_samples}{A matrix containing estimates from all bootstrap replications. Rows
represent bootstrap iterations, columns the term estimated.}
}
}
\description{
This function computes Wald and percentile bootstrapped confidence
intervals for the proposed VE estimator.
}
