% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_general_ci.R
\name{estimate_general_ci}
\alias{estimate_general_ci}
\title{Helper to estimate generic bootstrapped confidence interval}
\usage{
estimate_general_ci(
  one_boot_function_name,
  one_boot_args,
  ci_type,
  n_boot,
  pt_est = NULL,
  alpha = 0.05,
  return_boot = TRUE,
  n_cores = 1
)
}
\arguments{
\item{one_boot_function_name}{Character string naming the function that computes
one bootstrap iteration of estimates}

\item{one_boot_args}{List of arguments to pass to the function that computes
one bootstrap iteration of estimations}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{pt_est}{A matrix of estimates. The columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}

\item{n_cores}{Number of cores to use to when running bootstrapping procedure. Passed as \code{mc.cores} argument to \code{parallel::mclapply}.
If \code{n_cores > 1}, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).}
}
\value{
A list containing the following:
\describe{
\item{ci_estimates}{A list of matrices containing the lower and upper confidence interval bounds and
bootstrap standard error for each term. When \code{ci_type = "wald"}, the bootstrapped standard errors
on the transformed scale are also included.}
\item{n_success_boot}{A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are
those that result in non-missing valid point estimates.)}
\item{boot_samples}{If \code{return_boot = TRUE}, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and
the columns are the time points.}
}
}
\description{
This function computes Wald and percentile bootstrapped confidence
intervals for the proposed VE estimator.
}
