% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimates_by_time}
\alias{estimates_by_time}
\title{Reform estimates list to a list arranged by time point rather than term}
\usage{
estimates_by_time(estimates)
}
\arguments{
\item{estimates}{{A list of matrices for cumulative incidences and VE.
Each matrix contains the point estimate and confidence intervals for the specified term.}}
}
\value{
A list of matrices for each time point where the rows
are the term and the columns are the estimates at the given time point.
}
\description{
Reform estimates list to a list arranged by time point rather than term
}
