% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimates_to_df}
\alias{estimates_to_df}
\title{Convert estimates list to tidy data frame}
\usage{
estimates_to_df(estimates)
}
\arguments{
\item{estimates}{{A list of matrices for cumulative incidences and VE.
Each matrix contains the point estimate and confidence intervals for the specified term.}}
}
\value{
A data frame with columns \code{t0} specifying the time point and \code{term},
and point estimates, and confidence intervals.
}
\description{
Convert estimates list to tidy data frame
}
