% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_weights.R
\name{get_gp}
\alias{get_gp}
\title{Get empirical probability distributions g(d|x) and p(x) based on input data}
\usage{
get_gp(df, outcome_name, trt_name, time_name, adjust_vars, tau)
}
\arguments{
\item{df}{A data frame representing the target population for the estimated distributions
\itemize{
\item typically, population involves the "analysis-eligible" subset of a given data source
}}

\item{outcome_name}{Name of the possibly-censored time-to-event variable, measured from the chosen time origin.}

\item{trt_name}{Name of the exposure indicator. The underlying column should be numeric
(\code{1} = exposed during follow-up, \code{0} = never exposed during follow-up). Exposure indicator
should indicate whether an individual is exposed while uncensored and before experiencing the endpoint of interest.}

\item{time_name}{Name of the time to exposure, measured from the chosen time origin; use \code{NA} if not exposed.}

\item{adjust_vars}{Character vector of baseline covariates to adjust for when fitting the hazard models.}

\item{tau}{Numeric. Only vaccinated who are at-risk tau days after vaccination are included.}
}
\value{
A list of the following:
\describe{
\item{g_dist}{A data frame containing the covariate-conditional day probabilities.
Columns include \code{group_name} identifying the covariate group, \verb{<time_name> specifying the day, }prob\verb{specifying the covariate-conditional day probability, and each variables in}adjust_vars\verb{\} \\item\{p_dist\}\{A data frame containing the covariate probabilities. Columns include }group_name\verb{ identifying the covariate group,}prob\verb{specifying the covariate probability, and each variable in}adjust_vars`}
}
}
\description{
Get empirical probability distributions g(d|x) and p(x) based on input data
}
\keyword{internal}
