% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_distributions.R
\name{get_observed_gp}
\alias{get_observed_gp}
\alias{get_matching_gp}
\title{Estimate the marginalizing distribution in the observed or matched analysis-eligible
populations.}
\usage{
get_observed_gp(data, outcome_name, trt_name, time_name, adjust_vars, tau)

get_matching_gp(
  matched_adata,
  outcome_name,
  trt_name,
  time_name,
  adjust_vars,
  tau
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{matched_adata}{For \code{get_matching_gp}, a data frame representing the matched analysis-eligible population}
}
\value{
A list of two data frames containing the marginalizing distributions returned from calls to \code{\link[=get_gp]{get_gp()}}
}
\description{
\code{get_observed_gp} returns marginalizing distributions  based on the original observed data
\code{get_matching_gp} returns marginalizing distributions based on the matched-analysis data
}
