% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_matching_ve.R
\name{get_one_matching_ve}
\alias{get_one_matching_ve}
\title{Helper function to estimate marginal risk in matched data set}
\usage{
get_one_matching_ve(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = TRUE,
  return_models = TRUE
)
}
\arguments{
\item{matched_data}{A data frame representing a matched cohort}

\item{outcome_name}{Character string representing the original outcome variable in  \code{matched_data}}

\item{event_name}{Character string representing the original event variable in \code{matched_data}}

\item{trt_name}{Character string representing the original vaccination status variable in \code{matched_data}}

\item{time_name}{Character string representing the original vaccination time variable in \code{matched_data}}

\item{adjust}{A character vector  of variables in \code{adata} to adjust for, or a formula object.
Do not include treatment variable.}

\item{times}{A numeric vector of the time points at which to return cumulative incidence estimates}

\item{censor_time}{Time at which observations should be censored. Typically the \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately}

\item{return_models}{If \code{TRUE}, fitted survival mdoels are returned.}
}
\value{
The object returned by \code{\link[=estimate_marginal_risk]{estimate_marginal_risk()}}
}
\description{
First, creates the analysis matched data set. Then calls \code{\link[=estimate_marginal_risk]{estimate_marginal_risk()}},
providing the appropriate outcomes based on the analysis data set.
}
