% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_nomatch_ve_advanced.R
\name{get_one_nomatch_ve_advanced}
\alias{get_one_nomatch_ve_advanced}
\title{Internal function to compute cumulative incidence and effect measures}
\usage{
get_one_nomatch_ve_advanced(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  tau,
  times,
  censor_time = max(times),
  effect_measure = c("vaccine_effectiveness", "risk_ratio"),
  weighting = c("observed", "custom", "matched"),
  custom_weights = NULL,
  matched_dist_options = NULL,
  formula_0 = NULL,
  formula_1 = NULL,
  return_models = TRUE,
  return_gp_list = TRUE,
  return_matching = TRUE
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and any
variables listed in \code{covariates}.}

\item{tau}{Non-negative numeric value specifying the time after exposure that
should be excluded from the risk evaluation period. This argument is
primarily intended for vaccination exposures, where it is common to exclude
the time after vaccination when immunity is still building. Time must be
measured in the same units as that used for \code{outcome_time} and \code{exposure_time}
and should reflect the biological understanding of when vaccine-induced
immunity develops (usually 1-2 weeks). For non-vaccine exposures, \code{tau} can
be set to 0 (no delay period).}

\item{censor_time}{Time after exposure at which exposed
individuals are administratively censored during model fitting. Default:
\code{max(times)}. This limits estimation to the observed  period of interest and
prevents extrapolation beyond the largest evaluation time.
Typically, users should leave this at the default.}

\item{weighting}{Character string specifying how to construct marginalizing weights.
One of:
\itemize{
\item \code{"observed"} (default): estimate weights from the observed data;
\item \code{"custom"}: use user-specified weights provided via \code{custom_weights} argument;
\item \code{"matched"}: estimate weights from a matched cohort (for historical purposes).
}}

\item{custom_weights}{a \code{list(g_weights, p_weights)} providing weights for
marginalizing the time- and covariate-specific cumulative incidences. Must
have the following format:
\itemize{
\item \code{g_weights}: data frame with columns
\itemize{
\item all variables in \code{covariates}
\item \code{exposure_time} (time of exposure),
\item \code{prob} (probability of exposure at the given time within the covariate-group;
should sum to 1 within each covariate-group)
}
\item \code{p_weights}: data frame with columns
\itemize{
\item all variables in \code{covariates}
\item \code{prob} (probability of covariate-group; should sum to 1 over all covariate groups.)
}
}}

\item{matched_dist_options}{If \code{weighting = "matched"}, a list of
options created by \code{\link[=matched_dist]{matched_dist()}} controlling how the matched dataset
is used to estimate the marginalizing distributions.}

\item{formula_0}{Either a formula or a pre-fit \code{coxph} model object for the
unvaccinated hazard (original time scale). If a fitted model is provided,
it will be used directly instead of fitting a new model. Intended for advanced use.}

\item{formula_1}{Either a formula or a pre-fit  \code{coxph} model object for the
vaccinated hazard (time-since-exposure scale). If a fitted model is provided,
it will be used directly instead of fitting a new model. Intended for advanced use.}

\item{return_gp_list}{Logical; return marginalizing weights? Default is
TRUE.}

\item{return_matching}{Logical; return matched datasets? Default is
TRUE if \code{weighting = "matched"}. When \code{weighting != "matched"}, this is
automatically set to \code{FALSE}.}
}
\value{
List with components:
\describe{
\item{estimates}{List of matrices: \code{cuminc_0}, \code{cuminc_1}, \code{risk_ratio}, \code{vaccine_effectiveness}
where each matrix has one row per timepoint}
\item{model_0, model_1}{Fitted Cox models (if \code{return_models = TRUE})}
\item{gp_list}{Marginalizing distributions (if \code{return_gp_list = TRUE})}
\item{matched_data, matched_adata}{Matched datasets (if \code{weighting = "matched"})}
}
}
\description{
This is an internal function that performs the actual estimation
of cumulative incidences and derived effect measures
using the G-computation style approach. It is called by \code{\link[=nomatchVE]{nomatchVE()}} and \code{\link[=one_boot_nomatch]{one_boot_nomatch()}}.
Users should typically call these functions rather than calling this function directly.
For historical reasons, this function handles more complex inputs than
exposed in the \code{nomatchVE()} interface. In particular, it includes
an options to
\itemize{
\item set censor time in hazard model for the exposed
\item weight by the weights from a matched dataset and
\item provide hazard model formulas or prefit objects
}
}
\keyword{internal}
