% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_ve.R
\name{get_one_ve}
\alias{get_one_ve}
\title{Compute VE point estimate}
\usage{
get_one_ve(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  matched_dist_options,
  times,
  censor_time,
  tau,
  formula_0,
  formula_1,
  return_models = TRUE,
  return_gp_list = TRUE,
  return_matching = TRUE
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{marginalizing_dist}{Character string describing the type of estimated
day/covariate distributions to marginalize over. Values are "observed" or
"matched". Alternatively, for pre-specified marginalizing distributions can
be a list of 2 dataframes named \code{g_dist} and \code{p_dist}.}

\item{matched_dist_options}{If \code{marginalizing_dist == "matched"}, a list of
parameters controlling process of creating matched cohort and matched analysis
dataset from which marginalizing distributions will be estimated. List
can be created by \code{\link[=matched_dist]{matched_dist()}}.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{formula_0}{A formula or pre-fit model for estimating hazards in
unvaccinated group (recommended for use only by power users)}

\item{formula_1}{A formula or pre-fit model for estimating hazards in
vaccinated group. (recommended for use only by power users)}

\item{return_models}{Logical:  Should survival models be returned?}

\item{return_gp_list}{Logical: Should marginalizing distributions be returned?}

\item{return_matching}{Logical: Should matched datasets be returned? Default is
TRUE if marginalizing_dist = "matched" and FALSE otherwise.}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A matrix of estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.}
\item{model_0, model_1}{If \code{return_models = TRUE}, the fitted survival models
for unvaccinated and vaccinated.}
\item{gp_list}{If \code{return_gp_list = TRUE}, the list of marginalizing distributions used.}
\item{matched_data}{If \code{return_matching = TRUE}, the matched cohort}
\item{matched_adata}{If \code{return_matching = TRUE}, the analysis-eligible
matched cohort}
}
}
\description{
This function computes point estimates of vaccine effectiveness
based on the proposed method. Internally called by \link{obsve} and \link{one_boot_ve}.
}
