% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalize_psi_dx.R
\name{marginalize_psi_dx_t0}
\alias{marginalize_psi_dx_t0}
\title{Marginalize conditional cumulative incidences}
\usage{
marginalize_psi_dx_t0(psi_dx, gp_list = NULL, show_intermediates = FALSE)
}
\arguments{
\item{psi_dx}{A data frame with conditional risk predictions containing:
\itemize{
\item \verb{<covariates>}: all covariate columns
\item \verb{<exposure_time>}: vaccination day column
\item \code{psi_0_dx}: cumulative incidence under no vaccine,\eqn{\psi_0(t_0; d, x)}
\item \code{psi_1_dx}: cumulative incidence  under vaccine,\eqn{\psi_1(t_0; d, x)}
}

Typically output from \code{\link[=compute_psi_dx_t0]{compute_psi_dx_t0()}}.}

\item{gp_list}{A list of marginalizing weights with two components:
\describe{
\item{g_weights}{Data frame of covariate-conditional exposure-day probabilities \eqn{g(d \mid x)}.
Must include:
\itemize{
\item \code{group_id}:covariate group identifier
\item \verb{<exposure_time>}: exposure time variable
\item \code{prob_g}: probability of exposure time given the covariates
- all variables in \code{covariates}
}
\item{p_weights}{Data frame of covariate probabilities \eqn{p(x)}.
Must include:
- \code{group_id}: covariate group identifier
- \code{prob_p}: marginal probability of each covariate group
- all variables in \code{covariates}
}
}
}

Default is \code{NULL} in which case each row of \code{psi_dx} gets equal weight.}

\item{show_intermediates}{Logical that only applies when \code{gp_list} is not \verb{NULL; when }FALSE\verb{(default), the function performs marginalization in a single step and returns only the overall cumulative incidences \eqn{\bar{\psi}_0(t_0)} and \eqn{\bar{\psi}_1(t_0)}. When}TRUE``, performs the two-step marginalization
(first over days \eqn{d}, then covariate groups \eqn{x})
and returns intermediate group-level results \eqn{\psi_v(t_0; x)}.}
}
\value{
When \code{show_intermediates = FALSE}, a named numeric vector with two elements:
\itemize{
\item \code{cuminc_0}: Overall cumulative incidence under no exposure
\item \code{cuminc_1}: Overall cumulative incidence under exposure
}

When \code{show_intermediates = TRUE}, a list containing both overall results (as a vector)
and group-specific results (as a data frame).
}
\description{
Averages day- and covariate-specific cumulative incidences from \code{\link[=compute_psi_dx_t0]{compute_psi_dx_t0()}}
to produce overall cumulative incidences under each exposure. By default, performs
a simple average. If weights are provided for in \code{gp_list}, performs a weighted
average.
}
\seealso{
\itemize{
\item \code{\link[=compute_psi_dx_t0]{compute_psi_dx_t0()}} for generating \code{psi_dx} input
\item \code{\link[=canonicalize_weights]{canonicalize_weights()}} for creating \code{gp_list}
}
}
