% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_rolling_cohort.R
\name{match_rolling_cohort}
\alias{match_rolling_cohort}
\title{Match vaccinated and unvaccinated individuals using rolling cohort design}
\usage{
match_rolling_cohort(
  data,
  outcome_time,
  exposure,
  exposure_time,
  matching_vars,
  id_name,
  replace = FALSE,
  seed = NULL
)
}
\arguments{
\item{data}{Data frame with study population}

\item{outcome_time}{Name of the time-to-event/censoring variable. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured from the chosen time
origin; use \code{NA} if not exposed. Time must be measured in the same units
(e.g. days) as that used for  \code{outcome_time}.}

\item{matching_vars}{Character vector of variables to match on exactly}

\item{id_name}{Name of unique identifier variable of individuals}

\item{replace}{Logical. Allow controls to be reused? Default: \code{FALSE}}

\item{seed}{Integer for reproducibility. Default: \code{NULL}}
}
\value{
A list containing the following:
\describe{
\item{matched_data}{Data frame of matched pairs with original variables plus:
\itemize{
\item \code{match_index_time}: Matching time
\item \code{match_type}: "case" or "control"
\item \verb{match_<exposure>}: Treatment at matching
\item \code{match_id}: Pair identifier
}
}
\item{n_unmatched_cases}{Number of unmatched vaccinated individuals}
\item{discarded}{Logical vector indicating excluded individuals}
}
}
\description{
Creates 1:1 matched pairs of vaccinated ("cases") and
unvaccinated ("controls") individuals. Uses a rolling cohort design where
controls must be unvaccinated and event-free at the time they are matched to
a case.
}
\details{
For each vaccination time, newly vaccinated individuals are matched
to eligible controls using exact covariate matching. Controls are eligible if
unvaccinated and event-free at that time. Vaccinated individuals may appear
as a control (when they are not yet vaccinated) and as a case.
}
\examples{
matched_cohort <- match_rolling_cohort(
data = simdata,
outcome_time =  "Y",
exposure = "V",
exposure_time = "D_obs",
matching_vars = c("x1", "x2"),
id_name = "ID",
seed = 5678
)
}
