% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helpers.R
\name{matched_dist}
\alias{matched_dist}
\title{Parameters for creating a matched dataset and matched analysis data set}
\usage{
matched_dist(id_name = "ID", matched_data = NULL, replace = FALSE, seed = NULL)
}
\arguments{
\item{id_name}{Name of unique identifier variable of individuals}

\item{matched_data}{A data frame representing a matched cohort}

\item{replace}{Logical. Allow controls to be reused? Default: \code{FALSE}}

\item{seed}{Integer for reproducibility. Default: \code{NULL}}
}
\value{
A list with control parameters
}
\description{
Parameters for creating a matched dataset and matched analysis data set
}
\keyword{internal}
