% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_matching_ci.R
\name{one_boot_matching}
\alias{one_boot_matching}
\title{Compute one bootstrap replicate of matching-based VE point estimate}
\usage{
one_boot_matching(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  separate = FALSE,
  limit_type = "fixed",
  data = NULL,
  id_name = "ID",
  matching_vars = NULL,
  replace = FALSE,
  return_boot = TRUE
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort}

\item{outcome_name}{Name of the possibly-censored time-to-event variable, measured from the chosen time origin.}

\item{event_name}{Name of the event indicator. The underlying column should be numeric
(\code{1} = event, \code{0} = censored).}

\item{trt_name}{Name of the exposure indicator. The underlying column should be numeric
(\code{1} = exposed during follow-up, \code{0} = never exposed during follow-up). Exposure indicator
should indicate whether an individual is exposed while uncensored and before experiencing the endpoint of interest.}

\item{time_name}{Name of the time to exposure, measured from the chosen time origin; use \code{NA} if not exposed.}

\item{method}{Character string specifying method for survival estimation ("cox" for
Cox proportional hazards regression model or "km" for Kaplan-Meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{times}{Numeric vector of time after exposure at which to evaluate VE. All values must be greater than tau and
should reflect meaningful clinical timepoints (e.g., 30, 60, 90 days).}

\item{censor_time}{Time after exposure at which exposed
individuals are administratively censored during model fitting. Default:
\code{max(times)}. This is used to limit borrowing of information from beyond the time period of interest.}

\item{tau}{Time after exposure (vaccination)  to exclude as part of an "immune build-up" period. Must be non-negative.
Common values may include 0 days (no delay), 7 days (one week), 14 days (two weeks) and should
match the biological understanding of when vaccine-induced immunity develops.}

\item{separate}{If \code{method = cox} and \code{separate = TRUE}, Cox regression models are fit in the treated and untreated groups separately.}

\item{limit_type}{The default value is \code{limit_type = "fixed"}, wherein bootstrap samples are formed from the
original matched data set. If `limit_type = "limit", the matching procedure is performed in each
bootstrap iteration.}

\item{data}{If \code{limit_type = "limit"}, the original data used to constructed the matched cohort}

\item{id_name}{If \code{limit_type = "limit"}, character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{If \code{limit_type = "limit"}, a character vector containing the names of variables in \code{data} to match on.}

\item{replace}{If \code{limit_type = "limit"}, logical: Should matching be done with replacement?}

\item{return_boot}{Logical; return bootstrap draws in \code{boot_samples}?
Default: \code{TRUE}. Must be set to \code{TRUE} if user plans to use \code{\link[=add_simultaneous_ci]{add_simultaneous_ci()}}
to obtain simultaneous confidence intervals.}
}
\value{
A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.
}
\description{
Compute one bootstrap replicate of matching-based VE point estimate
}
\keyword{internal}
