% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nomatch_ci.R
\name{one_boot_nomatch}
\alias{one_boot_nomatch}
\title{Compute one bootstrap replicate of VE point estimate}
\usage{
one_boot_nomatch(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  weighting,
  times,
  censor_time,
  tau,
  boot_formula_0,
  boot_formula_1,
  matched_data,
  gp_list,
  limit_type
)
}
\arguments{
\item{data}{A data frame with one row per individual containing
the columns named in \code{outcome_name}, \code{event_name}, \code{trt_name}, \code{time_name},
and any variables listed in \code{adjust_vars}.}

\item{outcome_name}{Name of the possibly-censored time-to-event variable, measured from the chosen time origin.}

\item{event_name}{Name of the event indicator. The underlying column should be numeric
(\code{1} = event, \code{0} = censored).}

\item{trt_name}{Name of the exposure indicator. The underlying column should be numeric
(\code{1} = exposed during follow-up, \code{0} = never exposed during follow-up). Exposure indicator
should indicate whether an individual is exposed while uncensored and before experiencing the endpoint of interest.}

\item{time_name}{Name of the time to exposure, measured from the chosen time origin; use \code{NA} if not exposed.}

\item{adjust_vars}{Character vector of baseline covariates to adjust for when fitting the hazard models.}

\item{weighting}{Character string specifying the type of marginalizing weights to use.
Either:
\itemize{
\item \code{"observed"} (default): use the empirical distribution of time to exposure and covariates among the exposed as the
marginalizing weights. This provides close alignment with the weights implicitly used in matching.
\item \code{"custom"}: use user-specified weights provided in the \code{custom_weights} argument
}}

\item{times}{Numeric vector of time after exposure at which to evaluate VE. All values must be greater than tau and
should reflect meaningful clinical timepoints (e.g., 30, 60, 90 days).}

\item{censor_time}{Time after exposure at which exposed
individuals are administratively censored during model fitting. Default:
\code{max(times)}. This is used to limit borrowing of information from beyond the time period of interest.}

\item{tau}{Time after exposure (vaccination)  to exclude as part of an "immune build-up" period. Must be non-negative.
Common values may include 0 days (no delay), 7 days (one week), 14 days (two weeks) and should
match the biological understanding of when vaccine-induced immunity develops.}

\item{boot_formula_0}{A formula for estimating hazards in
unvaccinated group for bootstrap samples (not a model)}

\item{boot_formula_1}{A formula for estimating hazards in
vaccinated group for bootstrap samples (not a model)}

\item{matched_data}{A data frame representing the matched cohort}

\item{gp_list}{List of two dataframes named \code{g_dist} and \code{p_dist} representing
the marginalizing distributions to use.}
}
\value{
A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.
}
\description{
This function creates a bootstrapped sample and computes the
corresponding point estimate
}
\keyword{internal}
