% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ve_panel.R
\name{plot_ve_panel}
\alias{plot_ve_panel}
\title{Plot cumulative incidence/VE estimates}
\usage{
plot_ve_panel(object, ci_type)
}
\arguments{
\item{object}{An object returned from a call to \code{\link[=nomatchVE]{nomatchVE()}}, \code{\link[=matching_ve]{matching_ve()}} or \code{\link[=simultaneous_ci]{simultaneous_ci()}}}

\item{ci_type}{String indicating type of confidence interval to plot.
Takes values of  "wald", "percentile" or "simul".}
}
\value{
ggplot
}
\description{
Plot cumulative incidence/VE estimates
}
