% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_ci.R
\name{simultaneous_ci}
\alias{simultaneous_ci}
\title{Compute simultaneous confidence intervals for timepoints object}
\usage{
simultaneous_ci(object, alpha, seed = NULL)
}
\arguments{
\item{object}{Object created by call to \code{\link[=obsve]{obsve()}} or \code{\link[=matching_ve]{matching_ve()}}}

\item{alpha}{Significance level for confidence interval}

\item{seed}{Seed for reproducibility}
}
\value{
A dataframe with lower and upper simultaneous confidence intervals for
each timepoint
}
\description{
Compute simultaneous confidence intervals for timepoints object
}
