% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints}
\alias{timepoints}
\title{Evaluate VE at various time points of interest}
\usage{
timepoints(object, times)
}
\arguments{
\item{object}{An object returned from \code{obsve()}}

\item{times}{A numeric vector containing the times after vaccination at which to evaluate VE}
}
\value{
The original \code{obsve} object with the following changes/additions:
\describe{
\item{estimates}{A list of dataframe/matrix of the estimates at each timepoint}
\item{n_success_boot}{A vector of the number of bootstraps for each timepoint}
\item{boot_samples}{A list of matrices containing the bootstrap estimates for each timepoint}
\item{timepoints}{The timepoints at which VE was evaluated}
}
}
\description{
Evaluate VE at various time points of interest
}
