% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timepoints.R
\name{timepoints_to_df}
\alias{timepoints_to_df}
\title{Convert list of estimates to dataframe with t0 as column}
\usage{
timepoints_to_df(object)
}
\arguments{
\item{object}{Object created by timepoints}
}
\value{
Data frame with estimates across all time points
}
\description{
Convert list of estimates to dataframe with t0 as column
}
