% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ve.R
\name{add_ve}
\alias{add_ve}
\title{Calculate VE based on risk_0 and risk_1}
\usage{
add_ve(psi_bar)
}
\arguments{
\item{psi_bar}{{A named numeric vector containing point estimates for
\verb{risk_0, risk_1}}}
}
\value{
{A named numeric vector containing point estimates for
\verb{risk_0, risk_1}, and \code{ve}}
}
\description{
This function computes VE as the proportion reduction in risk due to
vaccination (ie. VE = 1 - risk_0 / risk_1)
}
