% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_matching_ve.R
\name{compute_km_ve}
\alias{compute_km_ve}
\title{Get the marginal cumulative incidence in the treated and untreated
groups based on Kaplan Meier estimation.}
\usage{
compute_km_ve(
  adata,
  adata_outcome_name,
  adata_event_name,
  adata_trt_name,
  eval_times,
  keep_models = TRUE
)
}
\arguments{
\item{adata}{A data frame that represents the analysis data set of a clinical trial.}

\item{adata_outcome_name}{Character string specifying the time to event variable in \code{adata}. The
time should be the time to event from vaccination/matched index date}

\item{adata_event_name}{Character string specifying the event variable in \code{adata}}

\item{adata_trt_name}{Character string specifying the treatment variable in \code{adata}}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure. All values must be greater
than \code{tau} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (tau+1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{keep_models}{Logical; return the two fitted hazard models used to compute
cumulative incidences?
Default: \code{TRUE}.}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A matrix of estimates. The columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.}
\item{models}{If \code{keep_models = TRUE}, the models used to compute risk and VE}
}
}
\description{
Get the marginal cumulative incidence in the treated and untreated
groups based on Kaplan Meier estimation.
}
\keyword{internal}
