% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_bar.R
\name{compute_psi_bar_times}
\alias{compute_psi_bar_times}
\title{Compute overall cumulative incidences at multiple timepoints}
\usage{
compute_psi_bar_times(
  fit_0,
  fit_1,
  exposure_time,
  eval_times,
  tau,
  newdata,
  gp_list
)
}
\arguments{
\item{fit_0}{A fitted model returned from \code{\link[=fit_model_0]{fit_model_0()}}}

\item{fit_1}{A fitted model returned from \code{\link[=fit_model_1]{fit_model_1()}}}

\item{exposure_time}{Name of the time-to-exposure variable  in \code{newdata}.
Used to compute \eqn{\psi_0(t_0; d,x)} where \eqn{d + \tau} and \eqn{d + t_0} are needed.}

\item{tau}{Delay period}

\item{newdata}{New data at which to do predictions.}
}
\value{
A matrix of estimates where the columns are the terms \code{cuminc_0} and \code{cuminc_1},
and the rows are the time points of interest.
}
\description{
Wrapper that computes cumulative incidences at multiple
timepoints. Calls \code{\link[=compute_psi_bar_t0]{compute_psi_bar_t0()}}  internally for each timepoint.
Models are fitted once before calling this function to allow efficient evaluation
at multiple timepoints without refitting.
}
\details{
This function expects models already fitted via \code{\link[=fit_model_0]{fit_model_0()}} and
\code{\link[=fit_model_1]{fit_model_1()}}. It performs G-computation by predicting and marginalizing over
conditional risks.
}
