% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_d.R
\name{compute_psi_d}
\alias{compute_psi_d}
\title{Compute paired day- and covariate- specific risks \eqn{\psi_0, \psi_1} at a specific time point}
\usage{
compute_psi_d(fit_0, fit_1, time_name, t0, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted model with a \code{predict()} method supporting \code{type = "survival"}
trained on \code{Surv(time, event) ~ X} with \emph{time from study start}.}

\item{fit_1}{A fitted model with a \code{predict()} method supporting \code{type = "survival"}
trained on \code{Surv(time, event) ~ D + X} with \emph{time since vaccination}.}

\item{time_name}{Name of the time variable column in \code{gp_list$g_dist} describing
when individuals are vaccinated. Used as the base time to which \code{tau}/\code{t0} are added to
compute the times that need to be evaluated for \code{predict_from_model_0}. Not
used in \code{predict_from_model_1} since predictions are evaluated at \code{tau}/\code{t0} directly.}

\item{t0}{Timepoint at which to evaluate cumulative incidence (relative to vaccination day)}

\item{tau}{Delay period}

\item{gp_list}{a \code{list(g_dist, p_dist)} providing weights for marginalization. This is
used so that predictions are only computed for (d,x) where g(d|x) > 0.
- \code{g_dist}: data frame with columns \code{time_name} (vaccination day),
all variables in \code{adjust_vars}, \code{group_name} (covariate-group id),
and \code{prob} (probability of day given group).
- \code{p_dist}: data frame with variables in \code{adjust_vars} and \code{prob}.
Currently only the \code{g_dist} element in \code{gp_list} is used.}
}
\value{
A data frame with one row per row of \code{gp_list$g_dist} containing:
identifiers/covariates and the columns \code{cuminc_0} and \code{cuminc_1}, as well as
supporting survival columns from both regimes.
}
\description{
#' @description
High-level wrapper that calls internal regime-specific predictors and returns
side-by-side conditional risks \eqn{\psi_0(t_0; d, x)} and \eqn{\psi_1(t_0; d, x)}
for each (d, x) in \code{gp_list$g_dist}.
}
\details{
\itemize{
\item \deqn{\psi_0(t_0; d, x) = 1 - S_0(d+t_0 \mid x)/S_0(d+\tau \mid x)}
\item \deqn{\psi_1(t_0; d, x) = 1 - S_1(t_0 \mid d, x)/S_1(\tau \mid d, x)}
}
}
\seealso{
\code{\link[=predict_from_model_0]{predict_from_model_0()}}
}
