% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ve.R
\name{compute_ve}
\alias{compute_ve}
\title{Compute VE at multiple timepoints}
\usage{
compute_ve(fit_0, fit_1, time_name, times, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted model with a \code{predict()} method supporting \code{type = "survival"}
trained on \code{Surv(time, event) ~ X} with \emph{time from study start}.}

\item{fit_1}{A fitted model with a \code{predict()} method supporting \code{type = "survival"}
trained on \code{Surv(time, event) ~ D + X} with \emph{time since vaccination}.}

\item{time_name}{Name of the time variable column in \code{gp_list$g_dist} describing
when individuals are vaccinated. Used as the base time to which \code{tau}/\code{t0} are added to
compute the times that need to be evaluated for \code{predict_from_model_0}. Not
used in \code{predict_from_model_1} since predictions are evaluated at \code{tau}/\code{t0} directly.}

\item{tau}{Delay period}

\item{gp_list}{a \code{list(g_dist, p_dist)} providing weights for marginalization. This is
used so that predictions are only computed for (d,x) where g(d|x) > 0.
- \code{g_dist}: data frame with columns \code{time_name} (vaccination day),
all variables in \code{adjust_vars}, \code{group_name} (covariate-group id),
and \code{prob} (probability of day given group).
- \code{p_dist}: data frame with variables in \code{adjust_vars} and \code{prob}.
Currently only the \code{g_dist} element in \code{gp_list} is used.}
}
\value{
A matrix of estimates where the columns are the terms \verb{cuminc_0, cuminc_1}, and \code{ve}
and the rows are the time points of interest.
}
\description{
#' Vectorized wrapper that computes VE and cumulative incidences at multiple
timepoints. Calls \code{\link[=compute_ve_t0]{compute_ve_t0()}}  internally for each timepoint.
Models are fitted once before calling this function to allow efficient evaluation
at multiple timepoints without refitting.
}
\details{
This function expects models already fitted via \code{\link[=fit_model_0]{fit_model_0()}} and
\code{\link[=fit_model_1]{fit_model_1()}}. It performs G-computation by predicting and marginalizing over
conditional risks. See \code{\link[=compute_ve_t0]{compute_ve_t0()}} for details.
}
