% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ve.R
\name{compute_ve_t0}
\alias{compute_ve_t0}
\title{Compute proposed causal VE estimand given the fitted survival models at a single
time points of interest}
\usage{
compute_ve_t0(fit_0, fit_1, time_name, t0, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted survival model for the unvaccinated group}

\item{fit_1}{A fitted survival model for the vaccinated group}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{t0}{A timepoint at which cumulative incidence should be evaluated}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{gp_list}{List of marginalizing distributions. Used to determine which x, d
will be used for prediction}
}
\value{
A named numeric vector containing point estimates for
\verb{psi_bar_0, psi_bar_1}, and \code{ve}
}
\description{
Compute proposed causal VE estimand given the fitted survival models at a single
time points of interest
}
