% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_matching_ci.R
\name{estimate_matching_ci}
\alias{estimate_matching_ci}
\title{Estimate bootstrapped confidence intervals for matching-based VE estimator}
\usage{
estimate_matching_ci(
  matched_data,
  outcome_name = outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = pair_censoring,
  separate = FALSE,
  ci_type = "wald",
  limit_type = "fixed",
  data = NULL,
  id_name = "ID",
  matching_vars = NULL,
  replace = FALSE,
  n_boot = 0,
  pt_est = NULL,
  alpha = 0.05,
  return_boot,
  n_cores = 1
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{method}{Character string specifying method for survival estimation ("cox" for
cox proportional hazards model or "km" for kaplan meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately.}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{limit_type}{If \verb{limit_type = "limit", the matching procedure is performed in each bootstrap iteration. If }limit_type = "fixed"`, bootstrap samples are formed from the
original matched data set.}

\item{data}{If \code{limit_type = "limit"}, the original data used to constructed the matched cohort}

\item{id_name}{If \code{limit_type = "limit"}, character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{If \code{limit_type = "limit"}, a character vector containing the names of variables in \code{data} to match on.}

\item{replace}{If \code{limit_type = "limit"}, logical: Should matching be done with replacement?}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{pt_est}{A matrix of estimates. The columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested timepoints for evaluation.}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}

\item{n_cores}{Number of cores to use to when running bootstrapping procedure. Passed as \code{mc.cores} argument to \code{parallel::mclapply}.
If \code{n_cores > 1}, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).}
}
\value{
A list containing the following:
\describe{
\item{ci_estimates}{A list of matrices containing the lower and upper confidence interval bounds and
bootstrap standard error for each term. When \code{ci_type = "wald"}, the bootstrapped standard errors
on the transformed scale are also included.}
\item{n_success_boot}{The number of bootstrap samples used to compute confidence interval}
\item{boot_samples}{A matrix containing estimates from all bootstrap replications. Rows
represent bootstrap iterations, columns the term estimated.}
}
}
\description{
Estimate bootstrapped confidence intervals for matching-based VE estimator
}
