% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_model_0}
\alias{fit_model_0}
\alias{fit_model_1}
\title{Fit survival models to estimate hazards and predict risk}
\usage{
fit_model_0(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  formula_0 = NULL
)

fit_model_1(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  t0,
  censor_time,
  tau,
  formula_1 = NULL
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{formula_0}{A formula or pre-fit model for estimating hazards in
unvaccinated group (recommended for use only by power users)}

\item{t0}{The time (relative to time of vaccination) at which to return
cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
If interested in multiple timepoints, recommended to set this to the maximum timepoint of interest
and then using \code{\link[=timepoints]{timepoints()}}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{formula_1}{A formula or pre-fit model for estimating hazards in
vaccinated group. (recommended for use only by power users)}
}
\value{
The fitted survival object
}
\description{
\code{fit_model_0()} fits a Cox model to estimate risk for unvaccinated individuals
\itemize{
\item this model includes all individuals, censoring vaccinated individuals
at their time of vaccination
\item TODO: need to censor individuals at d_max + t0? d_max not that well defined.
}

\code{fit_model_1()} fits a Cox model to estimate risk for vaccinated individuals
\itemize{
\item this model only includes vaccinated individuals at risk \code{tau} days after their vaccination
\item furthermore, individuals are censored at \code{t0} days after vaccination
}
}
