% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_distributions.R
\name{get_gp}
\alias{get_gp}
\title{Get empirical probability distributions g(d|x) and p(x) based on input data}
\usage{
get_gp(df, outcome_name, trt_name, time_name, adjust_vars, tau)
}
\arguments{
\item{df}{A data frame representing the target population for the estimated distributions
\itemize{
\item typically, population involves the "analysis-eligible" subset of a given data source
}}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}
}
\value{
A list of the following:
\describe{
\item{g_dist}{A data frame containing the covariate-conditional day probabilities.
Columns include \code{group_name} identifying the covariate group, \verb{<time_name> specifying the day, }prob\verb{specifying the covariate-conditional day probability, and each variables in}adjust_vars\verb{\} \\item\{p_dist\}\{A data frame containing the covariate probabilities. Columns include }group_name\verb{ identifying the covariate group,}prob\verb{specifying the covariate probability, and each variable in}adjust_vars`}
}
}
\description{
Get empirical probability distributions g(d|x) and p(x) based on input data
}
