% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_weights.R,
%   R/get_one_nomatch_ve_advanced.R
\name{get_observed_gp}
\alias{get_observed_gp}
\alias{get_matching_gp}
\title{Estimate the marginalizing distribution in the observed or matched analysis-eligible
populations.}
\usage{
get_observed_gp(data, outcome_time, exposure, exposure_time, covariates, tau)

get_matching_gp(
  matched_adata,
  outcome_name,
  trt_name,
  time_name,
  adjust_vars,
  tau
)
}
\arguments{
\item{outcome_time}{Name of the time-to-event/censoring variable. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured from the chosen time
origin; use \code{NA} if not exposed. Time must be measured in the same units
(e.g. days) as that used for  \code{outcome_time}.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. These covariates should include all known confounders of
exposure and censoring measured at the chosen time origin.}

\item{tau}{Numeric. Only vaccinated who are at-risk tau days after vaccination are included.}

\item{matched_adata}{For \code{get_matching_gp}, a data frame representing the matched analysis-eligible population}
}
\value{
A list of two data frames containing the marginalizing distributions returned from calls to \code{\link[=get_gp]{get_gp()}}
}
\description{
\code{get_observed_gp} returns marginalizing distributions  based on the original observed data
\code{get_matching_gp} returns marginalizing distributions based on the matched-analysis data
}
\keyword{internal}
