% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_weights.R
\name{get_observed_weights}
\alias{get_observed_weights}
\title{Estimate observed distributions of exposure eval_times and covariates}
\usage{
get_observed_weights(
  data,
  outcome_time,
  exposure,
  exposure_time,
  covariates,
  tau
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and any
variables listed in \code{covariates}.}

\item{outcome_time}{Name of the time-to-event/censoring variable. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured from the chosen time
origin; use \code{NA} if not exposed. Time must be measured in the same units
(e.g. days) as that used for  \code{outcome_time}.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. These covariates should include all known confounders of
exposure and censoring measured at the chosen time origin.}

\item{tau}{Non-negative numeric value specifying the time after exposure that
should be excluded from the risk evaluation period. This argument is
primarily intended for vaccination exposures, where it is common to exclude
the time after vaccination when immunity is still building. Time must be
measured in the same units as that used for \code{outcome_time} and \code{exposure_time}
and should reflect the biological understanding of when vaccine-induced
immunity develops (usually 1-2 weeks). For non-vaccine exposures, \code{tau} can
be set to 0 (no delay period).}
}
\value{
A list with two data frames:
\itemize{
\item \code{g_weights}: covariate-conditional exposure time  probabilities (\eqn{g(d \mid x)})
\item \code{p_weights}: covariate probabilities (\eqn{p(x)})

Each data frame includes all variables in \code{covariates}, and a \code{prob} column
with empirical probabilities.
\code{g_weights} additionally includes a \verb{<exposure_time>} column for exposure eval_times.
}
}
\description{
Computes two empirical probability distributions used to marginalize
time- and covariate- specific cumulative incidences:
\itemize{
\item \eqn{g(d \mid x)}: the distribution of exposure eval_times within each covariate subgroup
of exposed individuals who remain at risk \code{tau} days after exposure.
\item \eqn{p(x)}: the distribution of covariates among exposed individuals
who remain at risk \code{tau} days after exposure.
}

Called internally by \code{\link[=nomatchVE]{nomatchVE()}}. Provided as an example of the
structure needed if a user passes in \code{custom_weights}.
}
\examples{
weights <- get_observed_weights(simdata, "Y", "V", "D_obs",
                                   c("x1", "x2"), tau = 14)
str(weights)

}
