% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_matching_ve.R
\name{get_one_matching_ve}
\alias{get_one_matching_ve}
\title{Compute VE point estimate in matched data set}
\usage{
get_one_matching_ve(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = TRUE,
  return_models = TRUE
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{method}{Character string specifying method for survival estimation ("cox" for
cox proportional hazards model or "km" for kaplan meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately.}

\item{return_models}{Logical value: Should fitted survival models be returned?}
}
\value{
The object returned by \code{\link[=compute_marginal_ve]{compute_marginal_ve()}}
}
\description{
First, creates the analysis matched data set based on \code{matched_data}.
Then calls \code{\link[=compute_marginal_ve]{compute_marginal_ve()}} on the analysis data set, specifying
the appropriate outcome and treatment arguments.
}
