% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_helpers.R
\name{matched_dist}
\alias{matched_dist}
\title{Parameters for creating a matched dataset and matched analysis data set}
\usage{
matched_dist(
  id_name = "ID",
  matched_data = NULL,
  replace = FALSE,
  seed = NULL,
  pair_censoring = TRUE
)
}
\arguments{
\item{id_name}{Character string specifying name of a unique individual
identifier in \code{data}.}

\item{matched_data}{A data frame representing a matched cohort}

\item{replace}{Logical: Should matching be done with replacement?}

\item{seed}{Numeric: seed to set prior to performing random matching}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}
}
\value{
A list with control parameters
}
\description{
Parameters for creating a matched dataset and matched analysis data set
}
