% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_ve.R
\name{matching_ve}
\alias{matching_ve}
\title{Compute a matching-based estimator of VE with confidence intervals}
\usage{
matching_ve(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = FALSE,
  ci_type = "wald",
  limit_type = "fixed",
  data = NULL,
  id_name = NULL,
  matching_vars = NULL,
  n_boot = 0,
  alpha = 0.05,
  return_models = TRUE,
  return_boot = TRUE
)
}
\arguments{
\item{matched_data}{A data frame representing a matched cohort}

\item{outcome_name}{Character string representing the original outcome variable in  \code{matched_data}}

\item{event_name}{Character string representing the original event variable in \code{matched_data}}

\item{trt_name}{Character string representing the original vaccination status variable in \code{matched_data}}

\item{time_name}{Character string representing the original vaccination time variable in \code{matched_data}}

\item{adjust}{A character vector  of variables in \code{adata} to adjust for, or a formula object.
Do not include treatment variable.}

\item{times}{A numeric vector of the time points at which to return cumulative incidence estimates}

\item{censor_time}{Time at which observations should be censored. Typically the \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{limit_type}{Character string indicating whether the marginalizing
distributions of interest are the estimated distributions in the "limit"
or "fixed" sense. For a prespecified marginalizing_dist, limit_type is
always fixed.}

\item{data}{If limit_type = "limit", the original data used to constructed the matched cohort}

\item{id_name}{Character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{A character vector containing the names of variables in \code{data} to match on.}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{return_models}{If \code{TRUE}, fitted survival mdoels are returned.}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A list of matrices of the estimates at each timepoint. Rows of
each matrix are the terms "risk_0", "risk_1", "ve". Columns of each matrix
gives the point estimate and confidence intervals at the specified timepoint.}
\item{times}{The timepoints at which VE was evaluated}
\item{n_success_boot}{The number of bootstraps for all the timepoints. TODO: inconsistent with \code{\link[=timepoints]{timepoints()}}.}
\item{models}{If \code{return_models = TRUE}, the models used to fit the original point estimates}
\item{boot_samples_list}{If \code{return_boot = TRUE}, a list of matrices containing the bootstrap estimates for each timepoint}
}
}
\description{
This function is the main function for computing a matching-based estimator.
}
