% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_ve.R
\name{matching_ve}
\alias{matching_ve}
\title{Compute a matching-based estimator of VE with confidence intervals}
\usage{
matching_ve(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method = "km",
  adjust = NULL,
  pair_censoring = TRUE,
  separate = TRUE,
  times,
  censor_time,
  tau,
  ci_type = "wald",
  n_boot = 0,
  alpha = 0.05,
  limit_type = "fixed",
  data = NULL,
  id_name = "ID",
  matching_vars = NULL,
  replace = FALSE,
  return_models = TRUE,
  return_boot = TRUE,
  n_cores = 1
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{method}{Character string specifying method for survival estimation ("cox" for
Cox proportional hazards regression model or "km" for Kaplan-Meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{method = cox} and \code{separate = TRUE}, Cox regression models are fit in the treated and untreated groups separately.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{limit_type}{The default value is \code{limit_type = "fixed"}, wherein bootstrap samples are formed from the
original matched data set. If `limit_type = "limit", the matching procedure is performed in each
bootstrap iteration.}

\item{data}{If \code{limit_type = "limit"}, the original data used to constructed the matched cohort}

\item{id_name}{If \code{limit_type = "limit"}, character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{If \code{limit_type = "limit"}, a character vector containing the names of variables in \code{data} to match on.}

\item{replace}{If \code{limit_type = "limit"}, logical: Should matching be done with replacement?}

\item{return_models}{Logical value: Should fitted survival models be returned?}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}

\item{n_cores}{Number of cores to use to when running bootstrapping procedure. Passed as \code{mc.cores} argument to \code{parallel::mclapply}.
If \code{n_cores > 1}, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).}
}
\value{
A list containing the following:
\describe{
\item{estimates}{A list of matrices of the estimates at each timepoint. Rows of
each matrix are the terms "risk_0", "risk_1", "ve". Columns of each matrix
gives the point estimate and confidence intervals at the specified time point.}
\item{times}{The timepoints at which VE was evaluated}
\item{n_success_boot}{A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are
those that result in non-missing valid point estimates.)}
\item{boot_samples}{If \code{return_boot = TRUE}, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and
the columns are the time points.}
}
}
\description{
This function is the main function for computing a matching-based estimator.
}
