% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsve-package.R
\docType{package}
\name{nomatchVE-package}
\alias{nomatchVE-package}
\title{nomatchVE: Causal Estimand for Observational Vaccine Studies}
\description{
Functions to estimate vaccine efficacy in observational vaccine studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ewu16.github.io/nomatchVE/}
}

}
\author{
\strong{Maintainer}: Emily Wu \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
