% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsve-package.R
\docType{package}
\name{nomatchVE-package}
\alias{nomatchVE-package}
\title{nomatchVE: Estimate Causal Vaccine Effectiveness in Observational Studies Without Matching}
\description{
Implements estimators for a proposed causal estimand of vaccine effectiveness using observational data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ewu16.github.io/nomatchVE/}
  \item \url{https://github.com/ewu16/nomatchVE}
  \item Report bugs at \url{https://github.com/ewu16/nomatchVE/issues}
}

}
\author{
\strong{Maintainer}: Emily Wu \email{emily.wu2@emory.edu} (\href{https://orcid.org/0009-0003-5487-6753}{ORCID})

}
\keyword{internal}
