% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatchVE.R
\name{nomatchVE}
\alias{nomatchVE}
\title{Main function to estimate vaccine effectiveness using a G-computation
style estimator}
\usage{
nomatchVE(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  times,
  censor_time = max(times),
  tau = 14,
  ci_type = "wald",
  limit_type = "limit",
  n_boot = 0,
  alpha = 0.05,
  marginalizing_dist = "observed",
  matched_dist_options = NULL,
  formula_0 = NULL,
  formula_1 = NULL,
  return_models = TRUE,
  return_boot = TRUE,
  n_cores = 1
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{ci_type}{Character string indicating which type of confidence interval
to return ("wald", "percentile", "both")}

\item{limit_type}{Generally, should be set to "limit", the default value.  For historical purposes,
\code{limit_type} can also be set to fixed, which treats the marginalizing distributions as
fixed. For a prespecified marginalizing_dist, \code{limit_type} is always fixed.}

\item{n_boot}{Number of bootstrap replicated used to compute confidence
intervals}

\item{alpha}{Significance level used to compute confidence intervals.
Confidence intervals have nominal level \code{1 - alpha}.}

\item{marginalizing_dist}{Generally, should be set to "observed", the default value,
which estimates the marginalizing distributions based on the observed data.
For historical purposes, marginalizing_dist can also be set to "matched", which uses the marginalizing distributions
of a matched dataset and requires use of the \code{matched_dist_options} argument.
To providing a fixed choice of marginalizing weight functions, \code{marginalizing_dist}
should be a list of 2 dataframes named \code{g_dist} and \code{p_dist}. \code{g_dist} must be a dataframe with the columns
<time_name> describing the time of vaccination, all variables in \if{html}{\out{<adjust>}} , \code{group_name} identifying
each unique covariate group, and \code{prob} for the corresponding probability. \code{p_dist} must be a dataframe with
all variables in \if{html}{\out{<adjust>}} and variable \code{prob} for the corresponding probability.}

\item{matched_dist_options}{If \code{marginalizing_dist == "matched"}, a list of
parameters controlling process of creating matched cohort and matched analysis
dataset from which marginalizing distributions will be estimated. List
can be created by \code{\link[=matched_dist]{matched_dist()}}.}

\item{formula_0}{A formula or pre-fit model for estimating hazards in
unvaccinated group (recommended for use only by power users)}

\item{formula_1}{A formula or pre-fit model for estimating hazards in
vaccinated group. (recommended for use only by power users)}

\item{return_models}{Logical value: Should fitted survival models be returned?}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}

\item{n_cores}{Number of cores to use to when running bootstrapping procedure. Passed as \code{mc.cores} argument to \code{parallel::mclapply}.
If \code{n_cores > 1}, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).}
}
\value{
A list. In addition to information related to the call/arguments to \code{nomatchVE}, contains the following:
\describe{
\item{estimates}{A list of matrices for cumulative incidences and VE.
Each matrix contains the point estimate and confidence intervals for the specified term.}
\item{gp_list}{A list containing the distributions g(d|x) and p(x) used for marginalization}.
\item{model_0}{The fit object used to predict risk for unvaccinated group}
\item{model_1}{The fit object used to predict risk for vaccinated group}
\item{n_success_boot}{A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are
those that result in non-missing valid point estimates.)}
\item{boot_samples}{If \code{return_boot = TRUE}, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and
the columns are the time points.}
}
}
\description{
The primary function to estimate vaccine effectiveness.
}
