% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_ve.R
\name{one_boot_matching}
\alias{one_boot_matching}
\title{Compute one bootstrap replicate of matching-based VE point estimate}
\usage{
one_boot_matching(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = FALSE,
  limit_type = "fixed",
  data = NULL,
  id_name = NULL,
  matching_vars = NULL,
  return_boot = TRUE
)
}
\arguments{
\item{matched_data}{A data frame representing a matched cohort}

\item{outcome_name}{Character string representing the original outcome variable in  \code{matched_data}}

\item{event_name}{Character string representing the original event variable in \code{matched_data}}

\item{trt_name}{Character string representing the original vaccination status variable in \code{matched_data}}

\item{time_name}{Character string representing the original vaccination time variable in \code{matched_data}}

\item{adjust}{A character vector  of variables in \code{adata} to adjust for, or a formula object.
Do not include treatment variable.}

\item{times}{A numeric vector of the time points at which to return cumulative incidence estimates}

\item{censor_time}{Time at which observations should be censored. Typically the \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{TRUE}, models are fit in the treated and untreated groups separately}

\item{limit_type}{Character string indicating whether the marginalizing
distributions of interest are the estimated distributions in the "limit"
or "fixed" sense. For a prespecified marginalizing_dist, limit_type is
always fixed.}

\item{data}{If limit_type = "limit", the original data used to constructed the matched cohort}

\item{id_name}{Character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{A character vector containing the names of variables in \code{data} to match on.}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}
}
\value{
A list with three named elements, "risk_0", "risk_1", "ve". Each
element is a vector containing the estimates for the given term at each time point in \code{times}
}
\description{
Compute one bootstrap replicate of matching-based VE point estimate
}
