% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_matching_ci.R
\name{one_boot_matching}
\alias{one_boot_matching}
\title{Compute one bootstrap replicate of matching-based VE point estimate}
\usage{
one_boot_matching(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = FALSE,
  limit_type = "fixed",
  data = NULL,
  id_name = "ID",
  matching_vars = NULL,
  replace = FALSE,
  return_boot = TRUE
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{method}{Character string specifying method for survival estimation ("cox" for
Cox proportional hazards regression model or "km" for Kaplan-Meier estimation)}

\item{adjust}{If \code{method = "cox"}, a formula or  character vector containing the names of covariates in
\code{matched_data} to adjust for.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{pair_censoring}{Logical to indicate if matched pairs should be censored
when unvaccinated pair becomes vaccinated.}

\item{separate}{If \code{method = cox} and \code{separate = TRUE}, Cox regression models are fit in the treated and untreated groups separately.}

\item{limit_type}{The default value is \code{limit_type = "fixed"}, wherein bootstrap samples are formed from the
original matched data set. If `limit_type = "limit", the matching procedure is performed in each
bootstrap iteration.}

\item{data}{If \code{limit_type = "limit"}, the original data used to constructed the matched cohort}

\item{id_name}{If \code{limit_type = "limit"}, character string representing the individual identifier variable  in \code{data}}

\item{matching_vars}{If \code{limit_type = "limit"}, a character vector containing the names of variables in \code{data} to match on.}

\item{replace}{If \code{limit_type = "limit"}, logical: Should matching be done with replacement?}

\item{return_boot}{Logical value: Should bootstrap estimates be returned?}
}
\value{
A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.
}
\description{
Compute one bootstrap replicate of matching-based VE point estimate
}
