% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nomatch_ci.R
\name{one_boot_nomatch}
\alias{one_boot_nomatch}
\title{Compute one bootstrap replicate of VE point estimate}
\usage{
one_boot_nomatch(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  weights_source,
  eval_times,
  censor_time,
  tau,
  gp_list,
  limit_type
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and any
variables listed in \code{covariates}.}

\item{outcome_time}{Name of the time-to-event/censoring variable. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{outcome_status}{Name of the event indicator. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured from the chosen time
origin; use \code{NA} if not exposed. Time must be measured in the same units
(e.g. days) as that used for  \code{outcome_time}.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. These covariates should include all known confounders of
exposure and censoring measured at the chosen time origin.}

\item{weights_source}{Character string specifying the type of marginalizing weights
to use. Either:
\itemize{
\item \code{"observed"} (default): set the marginalizing weights to the empirical
distribution of exposure eval_times and covariates among the exposed. This
provides close alignment with the weights implicitly used in matching.
\item \code{"custom"}: use the user-specified weights provided in the \code{custom_weights} argument.
}}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure. All values must be greater
than \code{tau} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (tau+1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{tau}{Non-negative numeric value specifying the time after exposure that
should be excluded from the risk evaluation period. This argument is
primarily intended for vaccination exposures, where it is common to exclude
the time after vaccination when immunity is still building. Time must be
measured in the same units as that used for \code{outcome_time} and \code{exposure_time}
and should reflect the biological understanding of when vaccine-induced
immunity develops (usually 1-2 weeks). For non-vaccine exposures, \code{tau} can
be set to 0 (no delay period).}

\item{gp_list}{List of two dataframes named \code{g_weights} and \code{p_weights} representing
the marginalizing distributions to use.}
}
\value{
A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.
}
\description{
This function creates a bootstrapped sample and computes the
corresponding point estimate
}
\keyword{internal}
