% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_boot_nomatch.R
\name{one_boot_nomatch}
\alias{one_boot_nomatch}
\title{Compute one bootstrap replicate for G-computation method}
\usage{
one_boot_nomatch(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  eval_times,
  tau,
  custom_gp_list
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and any
variables listed in \code{covariates}.}

\item{outcome_time}{Name of the time-to-event/censoring variable. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{outcome_status}{Name of the event indicator. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured from the chosen time
origin; use \code{NA} if not exposed. Time must be measured in the same units
(e.g. days) as that used for  \code{outcome_time}.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. These covariates should include all known confounders of
exposure and censoring measured at the chosen time origin.}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure. All values must be greater
than \code{tau} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (tau+1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{tau}{Non-negative numeric value specifying the time after exposure that
should be excluded from the risk evaluation period. This argument is
primarily intended for vaccination exposures, where it is common to exclude
the time after vaccination when immunity is still building. Time must be
measured in the same units as that used for \code{outcome_time} and \code{exposure_time}
and should reflect the biological understanding of when vaccine-induced
immunity develops (usually 1-2 weeks). For non-vaccine exposures, \code{tau} can
be set to 0 (no delay period).}
}
\value{
The \code{pt_estimates} component returned by \code{\link[=get_one_nomatch_ve]{get_one_nomatch_ve()}},
for a bootstrap sample.
}
\description{
This function computes the point estimate for
a bootstrap replicate.
}
\keyword{internal}
