% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ci.R
\name{one_boot_ve}
\alias{one_boot_ve}
\title{Compute one bootstrap replicate of VE point estimate}
\usage{
one_boot_ve(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  times,
  censor_time,
  tau,
  boot_formula_0,
  boot_formula_1,
  matched_data,
  gp_list,
  limit_type
)
}
\arguments{
\item{data}{A data frame containing pertinent information for VE estimation.}

\item{outcome_name}{Character string specifying the name of the outcome
variable in \code{data}. Outcome variable should be a numeric representing follow-up time relative to study
start d0.}

\item{event_name}{Character string specifying the name of the event indicator
variable in \code{data}. Event indicator should be numeric-valued: 1=event, 0=censored.}

\item{trt_name}{Character string specifying name of the vaccination indicator
variable in \code{data}. Vaccination indicator is an indicator of ever receiving vaccine during the
follow-up period and should be numeric-valued: 1=yes, 0 = no.}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{adjust_vars}{A character vector containing the names of variables in
\code{data} to adjust for.}

\item{marginalizing_dist}{Generally, should be set to "observed", the default value,
which estimates the marginalizing distributions based on the observed data.
For historical purposes, marginalizing_dist can also be set to "matched", which uses the marginalizing distributions
of a matched dataset and requires use of the \code{matched_dist_options} argument.
To providing a fixed choice of marginalizing weight functions, \code{marginalizing_dist}
should be a list of 2 dataframes named \code{g_dist} and \code{p_dist}. \code{g_dist} must be a dataframe with the columns
<time_name> describing the time of vaccination, all variables in \if{html}{\out{<adjust>}} , \code{group_name} identifying
each unique covariate group, and \code{prob} for the corresponding probability. \code{p_dist} must be a dataframe with
all variables in \if{html}{\out{<adjust>}} and variable \code{prob} for the corresponding probability.}

\item{times}{A numeric vector containing the times (relative to time of vaccination) at which to return
the cumulative incidence estimates.}

\item{censor_time}{The time at which vaccinated individuals are censored during model fitting.
By default, this is set to \code{max(times)}.}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{boot_formula_0}{A formula for estimating hazards in
unvaccinated group for bootstrap samples (not a model)}

\item{boot_formula_1}{A formula for estimating hazards in
vaccinated group for bootstrap samples (not a model)}

\item{matched_data}{A data frame representing the matched cohort}

\item{gp_list}{List of two dataframes named \code{g_dist} and \code{p_dist} representing
the marginalizing distributions to use.}

\item{limit_type}{Generally, should be set to "limit", the default value.  For historical purposes,
\code{limit_type} can also be set to fixed, which treats the marginalizing distributions as
fixed. For a prespecified marginalizing_dist, \code{limit_type} is always fixed.}
}
\value{
A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative
incidence/VE terms and the rows are the requested time points for evaluation.
}
\description{
This function creates a bootstrapped sample and computes the
corresponding point estimate
}
