% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timepoints.R
\name{plot_ve_panel}
\alias{plot_ve_panel}
\title{Plot cumulative incidence/VE estimates}
\usage{
plot_ve_panel(plot_data, ci_type)
}
\arguments{
\item{plot_data}{A dataframe}

\item{ci_type}{String indicating type of confidence interval to plot.
Takes values of  "wald", "percentile" or "simul".}
}
\value{
ggplot
}
\description{
Plot cumulative incidence/VE estimates
}
