% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_d.R
\name{predict_from_model_0}
\alias{predict_from_model_0}
\alias{predict_from_model_1}
\title{Compute psi_0(t0; d, x) and psi_1(t0; d,x ), respectively.}
\usage{
predict_from_model_0(fit_0, time_name, t0, tau, gp_list)

predict_from_model_1(fit_1, time_name, t0, tau, gp_list)
}
\arguments{
\item{fit_0}{A fitted survival model for the unvaccinated group}

\item{time_name}{Character string specifying name of the vaccination time
variable in \code{data}. Vaccination time should be NA if never received vaccine
during follow-up.}

\item{t0}{A timepoint at which cumulative incidence should be evaluated}

\item{tau}{The time excluded after vaccination to allow building up of
immunity}

\item{gp_list}{List of marginalizing distributions. Used to determine which x, d
will be used for prediction}

\item{fit_1}{A fitted survival model for the vaccinated group}
}
\value{
A data frame of predictions for day and covariate groups in in \code{gp_list}
}
\description{
These functions compute the conditional day- and covariate-specific cumulative
incidences.
\code{predict_from_model_0()} calculates \verb{psi_0(t0; d, x)} by using x as covariates in prediction and obtaining survival
at d + t0 and d + tau days
\code{predict_from_model_1()} calculates \verb{psi_1(t0; d, x)} by using x and d as covariates in prediction
and obtaining survival at t0 days
}
